socket_clients/chatGPT/requirements.txt
openai
pyyaml
textwrap3
